feat: brand new UX/UI!
feat: new Segment filtering feature

fix(#209): Maintain IAM Role Chained Configurations based on SSO Assumer session

fix(#217): gpu_init on WSL2 - Leapp_0.8.1_amd64.deb

fix(#220): AWS Session Expiration Issue

+++ src/app/services/session/azure/azure.service.ts
@@ -60,22 +60,22 @@ export class AzureService extends SessionService {
     // extract accessToken corresponding to the specific tenant (if not present, require az login)
     let accessTokenExpirationTime;
     if(accessTokensFile) {
-      accessTokenExpirationTime = this.extractAccessTokenExpirationTime(accessTokensFile, (session as AzureSession).tenantId);
+      accessTokenExpirationTime = this.extractAccessTokenExpirationTime(accessTokensFile, (session as AzureSession).tenantId, (session as AzureSession).subscriptionId);
     }
 
     if (!accessTokenExpirationTime) {
       try {
         await this.executeService.execute(`az login --tenant ${(session as AzureSession).tenantId} 2>&1`);
         accessTokensFile = this.parseAccessTokens();
-        accessTokenExpirationTime = this.extractAccessTokenExpirationTime(accessTokensFile, (session as AzureSession).tenantId);
+        accessTokenExpirationTime = this.extractAccessTokenExpirationTime(accessTokensFile, (session as AzureSession).tenantId, (session as AzureSession).subscriptionId);
       } catch(err) {
         this.sessionDeactivated(sessionId);
         throw new LeappExecuteError(this, err.message);
       }
     }
 
-    // if access token is expired
-    if (new Date(accessTokenExpirationTime).getTime() < Date.now()) {
+    // if access token is expired OR doesn't exist (FOR VERSION >= 2.30.0)
+    if (!accessTokenExpirationTime || (new Date(accessTokenExpirationTime).getTime() < Date.now())) {
       try {
         await this.executeService.execute(`az account get-access-token --subscription ${(session as AzureSession).subscriptionId}`);
       } catch(err) {
@@ -90,7 +90,12 @@ export class AzureService extends SessionService {
       // az configure —default location <region(location)>
       await this.executeService.execute(`az configure --default location=${(session as AzureSession).region} 2>&1`);
       // delete refresh token from accessTokens
-      this.deleteRefreshToken();
+
+      //(FOR VERSION >= 2.30.0)
+      if(accessTokensFile) {
+        this.deleteRefreshToken();
+      }
+
     } catch(err) {
       this.sessionDeactivated(sessionId);
       throw new LeappExecuteError(this, err.message);
@@ -125,9 +130,25 @@ export class AzureService extends SessionService {
     }
   }
 
-  private extractAccessTokenExpirationTime(accessTokens: AzureSessionToken[], tenantId: string): string {
-    const correctToken = accessTokens.find(accessToken => accessToken._authority.split('/')[1] === tenantId);
-    return correctToken ? correctToken.expiresOn : undefined;
+  private async extractAccessTokenExpirationTime(accessTokens: AzureSessionToken[], tenantId: string, subscriptionId: string): Promise<string> {
+    if(accessTokens) {
+      /////////////////////////////////////////////////////////////////////////////////////////////////////
+      //!!! FOR AZURE CLI PRIOR TO 2.30.0 https://docs.microsoft.com/en-us/cli/azure/msal-based-azure-cli//
+      /////////////////////////////////////////////////////////////////////////////////////////////////////
+      const correctToken = accessTokens.find(accessToken => accessToken._authority.split('/')[1] === tenantId);
+      return correctToken ? correctToken.expiresOn : undefined;
+    } else {
+      ///////////////////////////////////////////////////////////////////////////////////////////////
+      //!!! FOR AZURE CLI >= 2.30.0 https://docs.microsoft.com/en-us/cli/azure/msal-based-azure-cli//
+      ///////////////////////////////////////////////////////////////////////////////////////////////
+      try {
+        const result = await this.executeService.execute(`az account get-access-token --subscription ${subscriptionId}`);
+        const correctToken = JSON.parse(result);
+        return correctToken ? correctToken.expiresOn : undefined;
+      } catch(err) {
+        return undefined;
+      }
+    }
   }
 
   private deleteRefreshToken(): void {
